+++ facebook/games/__init__.py
@@ -1,0 +1,2 @@
+""" This package contains all game related stuff. """
+
